Refactor range slider components into separate files
